feat: implement products CRUD

+++ backend/src/main/java/org/autoflex/web/dto/ProductionCapacityDTO.java
@@ -1,5 +1,9 @@
 package org.autoflex.web.dto;
 
+import jakarta.validation.constraints.DecimalMin;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+
 import java.math.BigDecimal;
 
 public class ProductionPlanItemDTO {
@@ -8,13 +12,13 @@ public class ProductionPlanItemDTO {
     public String productCode;
     public String productName;
     public BigDecimal unitPrice;
-    public Long producibleQuantity;
+    public BigDecimal producibleQuantity;
     public BigDecimal totalValue;
 
     public ProductionPlanItemDTO() {
     }
 
-    public ProductionPlanItemDTO(Long productId, String productCode, String productName, BigDecimal unitPrice, Long producibleQuantity, BigDecimal totalValue) {
+    public ProductionPlanItemDTO(Long productId, String productCode, String productName, BigDecimal unitPrice, BigDecimal producibleQuantity, BigDecimal totalValue) {
         this.productId = productId;
         this.productCode = productCode;
         this.productName = productName;
